requirements.lock.txt
numpy==2.2.6
opencv-python-headless==4.12.0.88
PyYAML==6.0.3
